Update container name and add new package dependency
Modified the container name from 'buzz_api' to 'buzz' in the docker-compose.yaml file. This aligns with the application specific naming conventions. Also, added a new dependency 'uber/h3-go/v4' version 'v4.1.0' in go.mod and go.sum files. This dependency is required for the new functionality being introduced in the server.