Add distance calculation functionality
Implemented a function that calculates the distance between two geographic coordinates using the Haversine formula in the utility package. This function is vital for calculating distances in geographic applications, for instance, in location-based services where distances between locations need to be determined. Additionally, a corresponding test function was created in the utils_test.go file to ensure its correct operation.