Add functionality to increment user's likes count
Updated the user model to include a likes_count field and implemented the functionality to increment it whenever user receives a like. This improvement provides tracking and visibility over number of likes a user earns, enhancing the social aspect of the application. Further updated related functions, tests and APIs to support this new feature.